order-service/requirements.txt
Flask==2.3.3
requests==2.31.0
python-consul==1.1.0